+++ lib/core/src/display.rs
@@ -1,4 +1,4 @@
-pub mod mesh_registry;
+pub mod camera;
 pub mod rendering;
 pub mod scene;
 pub mod symbol;
