--- lib/core/src/display/symbol/mod.rs
@@ -1,1 +1,0 @@
-pub mod display_object;
